+++ lib/src/features/app/ui/app_vm.g.dart
@@ -9,7 +9,7 @@ part of 'app_vm.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$AppVm on _AppVm, Store {
-  final _$themeModeAtom = Atom(name: '_AppVm.themeMode');
+  late final _$themeModeAtom = Atom(name: '_AppVm.themeMode', context: context);
 
   @override
   ThemeMode? get themeMode {