+++ lib/src/features/main/places/service/event/places_event.freezed.dart
@@ -17,6 +17,12 @@ final _privateConstructorUsedError = UnsupportedError(
 class _$PlacesEventTearOff {
   const _$PlacesEventTearOff();
 
+  PlacesEventGetPlace getPlace(int placeId) {
+    return PlacesEventGetPlace(
+      placeId,
+    );
+  }
+
   PlacesEventGetPlaces getPlaces(int page, String searchText) {
     return PlacesEventGetPlaces(
       page,
@@ -30,44 +36,43 @@ const $PlacesEvent = _$PlacesEventTearOff();
 
 /// @nodoc
 mixin _$PlacesEvent {
-  int get page => throw _privateConstructorUsedError;
-  String get searchText => throw _privateConstructorUsedError;
-
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
+    required TResult Function(int placeId) getPlace,
     required TResult Function(int page, String searchText) getPlaces,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
     TResult Function(int page, String searchText)? getPlaces,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
     TResult Function(int page, String searchText)? getPlaces,
     required TResult orElse(),
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
+    required TResult Function(PlacesEventGetPlace value) getPlace,
     required TResult Function(PlacesEventGetPlaces value) getPlaces,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
     TResult Function(PlacesEventGetPlaces value)? getPlaces,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
     TResult Function(PlacesEventGetPlaces value)? getPlaces,
     required TResult orElse(),
   }) =>
-      throw _privateConstructorUsedError;
-
-  @JsonKey(ignore: true)
-  $PlacesEventCopyWith<PlacesEvent> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
@@ -76,7 +81,6 @@ abstract class $PlacesEventCopyWith<$Res> {
   factory $PlacesEventCopyWith(
           PlacesEvent value, $Res Function(PlacesEvent) then) =
       _$PlacesEventCopyWithImpl<$Res>;
-  $Res call({int page, String searchText});
 }
 
 /// @nodoc
@@ -86,32 +90,147 @@ class _$PlacesEventCopyWithImpl<$Res> implements $PlacesEventCopyWith<$Res> {
   final PlacesEvent _value;
   // ignore: unused_field
   final $Res Function(PlacesEvent) _then;
+}
+
+/// @nodoc
+abstract class $PlacesEventGetPlaceCopyWith<$Res> {
+  factory $PlacesEventGetPlaceCopyWith(
+          PlacesEventGetPlace value, $Res Function(PlacesEventGetPlace) then) =
+      _$PlacesEventGetPlaceCopyWithImpl<$Res>;
+  $Res call({int placeId});
+}
+
+/// @nodoc
+class _$PlacesEventGetPlaceCopyWithImpl<$Res>
+    extends _$PlacesEventCopyWithImpl<$Res>
+    implements $PlacesEventGetPlaceCopyWith<$Res> {
+  _$PlacesEventGetPlaceCopyWithImpl(
+      PlacesEventGetPlace _value, $Res Function(PlacesEventGetPlace) _then)
+      : super(_value, (v) => _then(v as PlacesEventGetPlace));
+
+  @override
+  PlacesEventGetPlace get _value => super._value as PlacesEventGetPlace;
 
   @override
   $Res call({
-    Object? page = freezed,
-    Object? searchText = freezed,
+    Object? placeId = freezed,
   }) {
-    return _then(_value.copyWith(
-      page: page == freezed
-          ? _value.page
-          : page // ignore: cast_nullable_to_non_nullable
+    return _then(PlacesEventGetPlace(
+      placeId == freezed
+          ? _value.placeId
+          : placeId // ignore: cast_nullable_to_non_nullable
               as int,
-      searchText: searchText == freezed
-          ? _value.searchText
-          : searchText // ignore: cast_nullable_to_non_nullable
-              as String,
     ));
   }
 }
 
 /// @nodoc
-abstract class $PlacesEventGetPlacesCopyWith<$Res>
-    implements $PlacesEventCopyWith<$Res> {
+
+class _$PlacesEventGetPlace implements PlacesEventGetPlace {
+  const _$PlacesEventGetPlace(this.placeId);
+
+  @override
+  final int placeId;
+
+  @override
+  String toString() {
+    return 'PlacesEvent.getPlace(placeId: $placeId)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is PlacesEventGetPlace &&
+            const DeepCollectionEquality().equals(other.placeId, placeId));
+  }
+
+  @override
+  int get hashCode =>
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(placeId));
+
+  @JsonKey(ignore: true)
+  @override
+  $PlacesEventGetPlaceCopyWith<PlacesEventGetPlace> get copyWith =>
+      _$PlacesEventGetPlaceCopyWithImpl<PlacesEventGetPlace>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(int placeId) getPlace,
+    required TResult Function(int page, String searchText) getPlaces,
+  }) {
+    return getPlace(placeId);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
+    TResult Function(int page, String searchText)? getPlaces,
+  }) {
+    return getPlace?.call(placeId);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
+    TResult Function(int page, String searchText)? getPlaces,
+    required TResult orElse(),
+  }) {
+    if (getPlace != null) {
+      return getPlace(placeId);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(PlacesEventGetPlace value) getPlace,
+    required TResult Function(PlacesEventGetPlaces value) getPlaces,
+  }) {
+    return getPlace(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
+    TResult Function(PlacesEventGetPlaces value)? getPlaces,
+  }) {
+    return getPlace?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
+    TResult Function(PlacesEventGetPlaces value)? getPlaces,
+    required TResult orElse(),
+  }) {
+    if (getPlace != null) {
+      return getPlace(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class PlacesEventGetPlace implements PlacesEvent {
+  const factory PlacesEventGetPlace(int placeId) = _$PlacesEventGetPlace;
+
+  int get placeId;
+  @JsonKey(ignore: true)
+  $PlacesEventGetPlaceCopyWith<PlacesEventGetPlace> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $PlacesEventGetPlacesCopyWith<$Res> {
   factory $PlacesEventGetPlacesCopyWith(PlacesEventGetPlaces value,
           $Res Function(PlacesEventGetPlaces) then) =
       _$PlacesEventGetPlacesCopyWithImpl<$Res>;
-  @override
   $Res call({int page, String searchText});
 }
 
@@ -184,6 +303,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
+    required TResult Function(int placeId) getPlace,
     required TResult Function(int page, String searchText) getPlaces,
   }) {
     return getPlaces(page, searchText);
@@ -192,6 +312,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
     TResult Function(int page, String searchText)? getPlaces,
   }) {
     return getPlaces?.call(page, searchText);
@@ -200,6 +321,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
+    TResult Function(int placeId)? getPlace,
     TResult Function(int page, String searchText)? getPlaces,
     required TResult orElse(),
   }) {
@@ -212,6 +334,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
+    required TResult Function(PlacesEventGetPlace value) getPlace,
     required TResult Function(PlacesEventGetPlaces value) getPlaces,
   }) {
     return getPlaces(this);
@@ -220,6 +343,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
     TResult Function(PlacesEventGetPlaces value)? getPlaces,
   }) {
     return getPlaces?.call(this);
@@ -228,6 +352,7 @@ class _$PlacesEventGetPlaces implements PlacesEventGetPlaces {
   @override
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
+    TResult Function(PlacesEventGetPlace value)? getPlace,
     TResult Function(PlacesEventGetPlaces value)? getPlaces,
     required TResult orElse(),
   }) {
@@ -242,11 +367,8 @@ abstract class PlacesEventGetPlaces implements PlacesEvent {
   const factory PlacesEventGetPlaces(int page, String searchText) =
       _$PlacesEventGetPlaces;
 
-  @override
   int get page;
-  @override
   String get searchText;
-  @override
   @JsonKey(ignore: true)
   $PlacesEventGetPlacesCopyWith<PlacesEventGetPlaces> get copyWith =>
       throw _privateConstructorUsedError;