+++ lib/src/features/main/places/ui/place_card/place_card_vm.g.dart
@@ -9,7 +9,8 @@ part of 'place_card_vm.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$PlaceCardVm on _PlaceCardVm, Store {
-  final _$distanceAtom = Atom(name: '_PlaceCardVm.distance');
+  late final _$distanceAtom =
+      Atom(name: '_PlaceCardVm.distance', context: context);
 
   @override
   int? get distance {