+++ lib/src/features/place_detail/place_detail_vm.dart
@@ -1,9 +1,0 @@
-import 'package:mobx/mobx.dart';
-
-part 'place_detail_vm.g.dart';
-
-class PlaceDetailVm = _PlaceDetailVm with _$PlaceDetailVm;
-
-abstract class _PlaceDetailVm with Store {
-  _PlaceDetailVm();
-}