+++ lib/src/navigation/app_router.dart
@@ -3,8 +3,8 @@ import 'package:where_to_go_today/src/features/auth/code/code_route.dart';
 import 'package:where_to_go_today/src/features/auth/register/register_route.dart';
 import 'package:where_to_go_today/src/features/auth/sign_in/sign_in_route.dart';
 import 'package:where_to_go_today/src/features/main/main_route.dart';
+import 'package:where_to_go_today/src/features/main/places/ui/place/place_route.dart';
 import 'package:where_to_go_today/src/features/onboard/ui/onboarding_route.dart';
-import 'package:where_to_go_today/src/features/place_detail/place_detail_route.dart';
 import 'package:where_to_go_today/src/features/settings/ui/settings_route.dart';
 
 /// Класс, в котором описываем навигацию в приложении.
@@ -25,8 +25,8 @@ class AppRouter {
       SettingsRoute.routeName: (_) => SettingsRoute(),
       '${MainRoute.routeName}/:initialScreen': (routeData) =>
           MainRoute(routeData.pathParameters['initialScreen']),
-      '${PlaceDetailRoute.routeName}/:placeId': (routeData) =>
-          PlaceDetailRoute(routeData.pathParameters['placeId'] ?? ''),
+      '${PlaceRoute.routeName}/:placeId': (routeData) =>
+          PlaceRoute(routeData.pathParameters['placeId'] ?? ''),
     },
   );
 }