+++ lib/src/res/asset.dart
@@ -5,7 +5,12 @@ class Asset {
 
 class _Png {
   static const _path = 'assets/images/3.0x/';
+  static const _pathPlace = 'assets/images/place/';
+
   final logoWtgt = '${_path}wtgt_logo.png';
+  final placeImage = '${_pathPlace}place.png';
+
+  final iconCocktail = '${_pathPlace}icon_cocktail.png';
 }
 
 class _Svg {